Break delaunator into pieces (#2808)
* Update delaunator to fix ODR issue with multiple headers.

* Use the standard delaunator header.

* Add missing headers.
