+++ filters/HAGFilter.cpp
@@ -36,7 +36,7 @@
 
 #include <pdal/KDIndex.hpp>
 
-#include "private/delaunator-decl.hpp"
+#include "private/delaunator.hpp"
 
 #include <string>
 #include <vector>
